add write and append methods

+++ log_data_for_chart.py
@@ -1,7 +1,7 @@
 from sensor_data import Sensor
 from csv import DictWriter
 from datetime import datetime
-
+from os.path import isfile
 
 def get_data():
     '''
@@ -22,15 +22,23 @@ def get_data():
 
 def write_data(data):
     """
-    Takes in data and writes the result into a CSV file. 
+    Takes in data and writes the result into a CSV file.
+    If the file does not exist(has been deleted), then create a new file, write a header and append the data. 
+    Else, the CSV file already exists(has not yet been deleted), then just append data.
     """
     file_name = "data_log.csv"
-    with open(file_name, "w") as csv_file:
-        headers = ["temperature", "humidity", "pressure", "date"]
-        csv_writer = DictWriter(csv_file, fieldnames=headers)
-        csv_writer.writeheader()
-        csv_writer.writerow(data)
-
-
+    headers = ["temperature", "humidity", "pressure", "date"]
+    mode = "a" if isfile(isfile(f"./{file_name}")) else "w"
+    
+    if(not isfile(f"./{file_name}")):
+        with open(file_name, "w") as csv_file:
+            csv_writer = DictWriter(csv_file, fieldnames=headers)
+            csv_writer.writeheader()
+            csv_writer.writerow(data)
+    else:
+        with open(file_name, "a") as csv_file:
+            csv_writer = DictWriter(csv_file, fieldnames=headers)
+            csv_writer.writerow(data)            
+            
 data = get_data()
 write_data(data)
